walking works
vr walking works
no in game gui though

+++ src/main/java/com/example/examplemod/PlayerExtension.java
@@ -27,8 +27,8 @@ public interface PlayerExtension {
 	void swingArm(InteractionHand interactionhand, VRFirstPersonArmSwing interact);
 
 	boolean isClimbeyJumpEquipped();
-	double getMuhJumpFactor();
-	double getMuhSpeedFactor();
+	float getMuhJumpFactor();
+	float getMuhSpeedFactor();
 	double getRoomYOffsetFromPose();
 	boolean getInitFromServer();
